feat: PrivateKeySigner
